requirements.txt
absl-py==0.3.0
astor==0.7.1
click==6.7
flake8==3.5.0
Flask==1.0.2
Flask-WTF==0.14.2
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
itsdangerous==0.24
Jinja2==2.10
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.5
pandas==0.23.4
Pillow==5.2.0
protobuf==3.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
Werkzeug==0.14.1
WTForms==2.2.1
